"9:30pm. Whops. I just realized I made a serious mistake when inferring union patterns.
```fs
                | ValueNone ->
                    match Map.tryFind name aux.union_cases with
                    | Some i -> assume i
                    | None -> errors.Add(r,CasePatternNotFound name); f (fresh_var()) a
```

```fs
type AuxEnv = {
    hoc : Map<int,HigherOrderCases>
    union_cases : Map<string,int>
    }
```

Ah, crap. Those `union_cases` are a huge mistake. It is because they do not obey opens.

Instead of doing this, I should be looking into the current (or maybe just the top level) environment and then ripping the type straight from the higher order dictionary.

```fs
match Map.tryFind name top_env.ty with
```

Yeah, for both of these, I should be looking into the current environment. I had not been thinking opens at all when doing this part. It won't be too hard to fix, but I am too tired to do it right now. I'll get back to this when I am done with the editor support stuff.

9:40pm. It is really easy to miss obvious stuff like this when you are overwhelmed. This kind of mistake is due to my inexperience. This is my first time implementing a top down type system using unification, that much is obvious.

9:45pm. I would not dare to run the typechecker directly right now even if I could. I need to get the scaffolding up first. Then I'll go through all the cases one by one once again and fill out the missing pieces, testing everytihng as I go along. Right now I think the typechecker is 80-90% complete which is very good. All the essentials are there. The completed form is in sight. All I have to do is put in another 3 weeks of effort into getting it all to work.

Then comes the prepass. And the partial evaluator. And after that I will just plug in the code gen from the prototype and at that point I will be able to program for real in v0.2.

Damn it, this is just 2017 all over again, except this time I am making Spiral to overcome issues caused by my previous design. It was too much to expect at that time to be done with making languages. Even if it wasn't for a the top down type system, v0.09 had too many scalability issues. This time I will get it all done properly.

I really hope that in 2021, I will put language work behind me and will be able to do what I really want which is ML.

If something does not work at first, try it again with 1,000x more power."