"10:20am. I went to bed early, but I got up late. But at least I slept well and am not as groggy as usual.
Let me chill for a while and then I will start. New Ass Pride is out.

No matter how I look at it, my programming journey is like a nightmare. I am doing unpaid labor for some far off goal, just spending my life for it. But it does need to be done.

I will only be free once I am done with Spiral. I might have finished the typechecker, but it is not the time to relax yet. I released my tension too soon. Burden and pressure is my present.

I am really unlucky. It is a cruel and unusual punishment that all the power I am getting from making Spiral has no bearing upon the real world. And even if I win, the real world will only be about programming. Too bad about the NPCs. I regret it, but it is not my fault. I only saw the future and made to grasp it. I am not the one who created this universe. I deserve none of the blame for the state of the affairs that will come.

We will all get what is coming to us.

So I must take courage. At this point I am at least halfway done with Spiral v0.2. I can afford to keep going for another two months, that is for sure. I will finish the language and then leave it up to fate. If I can get my hands on those chips, that is fine. In that case I will continue with my plan of continuing to build my programming skills. I'll try to make that poker agent. If not, I'll take a much deserved break from all of that and focus on money just for a bit.

10:30am. The real world is so unfair. Why is it the normies who get all the smooth gradients while I toil away in irrelevance?

Well, I am smart so I know why - because they collectively create the outside. The outside is not nature - it is a bubble created by them. If I want a world where people like me can prosper words and theories are not enough. I need a power strong enough to burst it beyond repair.

10:35am. I can wail at this, but I must always remember that power is worth it for it own sake. My past is nothing but a distraction. The rational agent would just act with his future best interests in mind. Morality should be subservient to that.

Compare programming now to programming of 30 years ago. And imagine what it would be 30 years hence even if it wasn't for the Singularity. It will be something completely different. I need to grasp some of that benefit today.

And for that I need Spiral.

I need to eke out just this single edge. In 2016 I really developed my current style of programming, in 2018 I rebased it so it includes staging, and now I will perfect it.

With Spiral v0.2, no matter what kind of hardware comes out, I will be able to have a perfect handle on it. I proved that I can do this with GPUs. I just need to go a little bit further.

10:45am. Back in late 2018 I had essentially hit a wall in terms of productivity. It was not really obvious in April 2018 when I had really started using Spiral for serious things, but six months later, I was sick of how much the language was making me work. I could not bear it.

There is a certain kind of self pity that comes from that. All the wins were did not matter, but all the losses did. I could definitely look at Deepmind and OpenAI guys with barely constrained envy.

10:50am. I guess this is what I will be doing today - just ranting. My mind is weak and I need to fortify my drive. I am not a machine. Even though I know what my purpose in life is, and what I should be doing right now, sometimes the tedium of it overcomes me.

This is the greatest insult of all to me. If it were the Inspired, they would always have the strength to believe in their purpose. But inefficient as it is, feeling bad about not fulfilling my goals will get me back to them.

Oh yes, one of the reasons I am holding off on resuming Simulacrum is because seeing characters believe in their purpose is just too painful for me right now. I can barely constrain my envy for fictional things.

11:05am. There isn't a person that I know of exists that I look up to, but if somebody could do all that I am doing now without regret and with supreme fervor, for once I would accept such a person as my superior.

Ahhhh...the Inspired Desire. It is so near and yet so far away. It is beyond a wall made of glass. I can see it, but I cannot grasp it.

11:10am. Let me commit here, and finally actually do my chilling instead of just brooding. A bunch of things are out, and when I get back, I think I will just spend the day in bed until I get the feeling to do it. Eventually, as I imagine doing editor support, the arduous labor I did on the typechecker will become a distant memory and I will break my shackles. I need to warm up towards the typechecking server.

The best way to do it is to keep rolling it around in my mind."