"12pm. It is important to enjoy the struggle in life.
When you are strong, the challenges seem like good fun. The life of battling demons in hell seems like entertainment.

When you are weak, it feels like you are going to die tomorrow and yet there is so much work left to be done. You forget that the struggles were once enjoyable. You can't stop thinking about the past.

12:05pm. The work on the typechecker just blew through my mental reserves.

Right now, what I need the most aren't necessarily schematics for how to do editor support. Rather it is hope. With enough strength, the despair becomes hope itself. With enough strength, hell becomes heaven.

12:10pm. I need to remember the feeling of power.

1:15pm. Done with breakfast. Let me read the two Ancestor Is Acting Like A Loli chapters and I will go to bed. To be honest, I do not feel like reading even this. Yeah, actually forget it. Let me just go to bed, and I'll try to grab on to that feeling of confidence. Forget useless fiction.

The story being carved out in one's own mind is always the best story. I just have to make up my mind to impose it upon reality. The inner conflicts need to be resolved.

Let me shut down for a while. This is what I want to do.

Until Spiral is done, all fiction is useless. Programming is what I should want to be doing. I tread it as a burden, but this is the only thing in life worth doing. I need to remind myself of this."