"6:25pm. I've been in bed long. Yeah, these mediation sessions do wonders for my will.
Today's main theme was definitely wizards. I've been thinking about wizards and hard work.

To me, hard work was always something to be avoided. I always considered hard work to be a mistake. And with regards to money, I would only consider it validly made if it was easy. The main reason I really failed at trading is because I could never allow myself to get a job so I could get a stake.

Money...is not an actual power, anymore than social status is. Sure it is a half power, but it is a means, not an actual end. No matter how much money you have, you will never have more than the guys with the authority to just print it. And social status to the weak willed might seem everything, but how do you know you aren't some monkey competing for prestige amongst other monkeys?

If I went in Africa to live in the jungles, I would not have the urge to be respected by various species of animals living there.

6:55pm. I do not necessarily see the others around me as animals, but if you look at history, you get the sense that believing society and culture is not the best thing to do. I do not really get the sense that people in general undertand their place in history. I do look down on them for that. They have too careless. They are too willing to blindly believe the force known as culture.

I haven't done this right. Though I took a step away from society, I had not internalized the right values.

Take the Singularity for instance. It will most likely happen late in this decade, and if not, Kurzweil's timeline should hold regardless.

But that is not what I should believe in because such a belief won't matter either way.

What I need to believe in are things that if it does happen according to schedule will let me take advantage of it.

I need to aim for wizardry.

A good story would be one of a mediocre wizard mastering one spell tier every decade. I get it - it would be better if the wizard mastered a spell tier every few years. It would be even better to master it in just a year or every few months. It would be better if he was a genius.

But such a thing is not 'realistic'.

Really, the one thing chinks get right in their stories is respect for elders. Where else, but in cultivation novels do you get old people actually being useful? It is a reflection of their real world attitudes.

Now, actual real world old people are pretty useless, but training people to respect them is probably better than training people to pray to truck-kun for help. Cultivating for decades or studying magic, either is better than hoping for a chance in your next life.

Life is long. I can't live it thinking I am going to upload myself in less than a decade. Or that I will be able to make Skynet in that time.

The ideas I've come up with are ultimately cheap. The one who wins will be the one who really enjoys programming.

7:20pm. I want to program. I honestly feel like that. But I do not want to be rushed. It is fun to design things in your head first. Trying to force that is the opposite.

Here I am having trouble starting editor support. I want to finish Spiral as soon as possible, like in the next few months so I can get on with 'bigger and better' things.

But really, work on Spiral is the work on my foundation. Even if I become some godlike being in the future, literally nothing will be more important than the work I am doing right now.

I should be just fine letting it take longer. I can afford to play for a year or two. That is how thing will go. Yes, in a few months I will have it in a usable state, but it will take me a while longer to make it really polished.

Money or respect does not matter. When you start whining about that, you know you've failed at life.

7:25pm. I'll let this part take as long as needed. Once I am able to cast spells, I'll go to town and sell a few scrolls and magic items of my making. That is the way things go.

Instead of building my power properly, I wasted time in school and then wasted some more trading after that. Late starts, lack of talen, those are all disadvantages.

But the Singularity might take a while, and I might end up living surprisingly long. My grandfaster lived up to his 80s I think. So that means roughly that I have 50 years in which to do programming left.

I have to be like a wizard. Master the foundations, master low level spells, master the mid ones and so on.

Personal power is what is really important in life. Or rather, the effort taken into building it is. The belief in that effort is the most important thing.

7:30pm. Technology is always a play with power you do not understand. True understanding would be the ability to see everything. What humans call understanding is merely having a good story. Up to this day, nobody has every explained inspiration.

Thus far, the benefits of technology have been directed to the individual through society.

But 'society' is not the one working on programming languages. It is not the one trying to master new techniques. Society is not the one trying to transform itself. Society is only willing to pay lip service to progress, it does not have what it takes to harness it.

I want programming one day to be as powerful as real magic. I need to be willing to work for this goal.

Isn't the real reason I've started so late on it is because it is so looked down on? Even though I had talent, I knew pretty well that being able to program in Basic and Logo is quite different from being able to cast fireball.

7:40pm. I won't let this stand.

Would it not be great if somebody came along and changed the status quo.

Sure, normies would get fucked in such an environment, but one can consider that as a side benefit. If they cannot contain the flame, they should not be playing with sparks.

7:45pm. A good philosophy should be all about personal effort. If one does not desire to work, then one should first desire to have that desire.

Was doing the typechecker that strenous?

Did I not learn how unification works? Did I not get something out of that effort?

I was really much more patient during the first six months than now.

7:50pm. Work is a nightmare. Especially during the low power segments like now. But even then, if you accept that curse, you become free from other chains.

Like for example, had I the resolve to do what I am doing now, I could have skipped high school entirely.

I had not really changed during the time, so all that decision would have done is bring me destiny 12 years closer.

7:55pm. Right now programming is just an excuse, but is stopping me from having to stock shelves, or some other kind of meaningless work.

...Yeah, I can't forget that either. I should remind myself of what the alternatives are to being a wizard every time I do not feel like it.

8pm. Let me stop here. Maybe I will meditate on it tomorrow as well. Right now fatigue is hitting me like a ton of bricks. Churning my brain at max capacity while standing still is very mentally draining.

Either I will spend one more day like this, or I will reassert my determination tomorrow and get back to my usual pace.

So far, my mode of operation was to get every advantage available before I start. I am a complete madman who has done that for over 5 years and counting. A few years more won't hurt before I graduate to the next level.

One last thing, before I close. If Spiral is the foundation, that does mean that the others are operating with broken ones. In this competition against the world, I won't let the chance slip. Either way, I will get what is coming to me.

My language will be supreme."