"10:30am. I am up. Let me chill for a while and I will try to get some work done. I've been reading manga till late, but on the plus side I slept well.
I need to make up my mind. Either I get back to leveling up my programming skills, or I pretty much stay a loser for the rest of my life. I can either get to work, or wallow in my regret.

When it comes to doing intelligence augmentation for humans, the best way to do it for a human is to just let the machine do the task. You do not get anything better than that. So if I want to get smarter, this is what I should be practicing.

Necromutation is really a high level spell. A sensible wizard would start off by desiring something smaller.

10:35am. I myself, even though I am trying to be rational about it have too much of a shot for the stars mentality. It is really amazing when you think about it that I spent all this time on the foundations in spite of that. For some reason the people working sensibly are not that interested in it.

It is a clown world. None of the hierarchies make a lick of sense in it. Weak is strong, and I am perenially confused.

10:40am. Forget it. There is no need to accept the Singularity. The path towards inhumanity should be gradual.

I could not accept becoming a programmer back then because I did not see it as power, but what did I expect? That I should have direct access to my own brain? That would have just gotten me killed.

Before I master editing myself, how about I master editing others first? Much safer to do it this way.

10:45am. It is this anti-social tendency to not feel like the power is yours unless you can hold it, that has been killing me from day one. If I had this belief, I would have been so much stronger.

I should be aiming to be a wizard. Unlike dumb peasants shoveling dirt, they only get stronger with age. At the age of 80 peasants are getting ready to be put in a casket, but wizards are readying their undead loli transformation.

10:55am. Let me do my morning reading and I will start. I am not too inspired, so I think once it is time to do so that I will just look at the code that I've written previously."