"12:40pm. Let me finally start. The long term might be long, but will also soon be here.
Again and again it gets tedious, and every time I will reinvent myself to bind myself more tightly to the work that I do. My programming should not be outside me.

12:45pm. For a while from here on out, I will get rid of all the distractions. Every time I start reading stuff on the side, it eats away my motivation to do programming. What desire I build up storming, it gets frittered away as soon as I indulge my hobbies.

12:55pm. Ok...I am thinking. It is ironic of all the problems I have to deal with, that my focus now is entirely on the hover provider.

Let me describe the problem.

You have a piece of code, you hover the cursor over a variable, then you wait for the type info to appear. The hover provider is what sends a message to the server to get this information.

The problem is - what if typechecking takes a long time? In that time the user could move the cursor somewhere else.

But with distributed communication that I am doing it is not easy to tell the server that this has occurred. On the VS Code side it is expecting a promise. And the F# side, it will be waiting for the typechecking to finish.

Even if I used a router socket, I can't actually switch cleanly because Hopac is not integrated with NetMQ. I can't change hovers until the current one finishes.

```fs
        match x with
        | ProjectFileOpen x ->
            match config x.spiprojDir x.spiprojText with Ok x -> [||] | Error x -> x
            |> Json.serialize
        | FileOpen x ->
            (let res = IVar() in Ch.give (server_blockizer x.spiPath) (Req.Put(x.spiText,res)) >>=. IVar.read res)
            |> run |> Json.serialize
        | FileChanged x ->
            (let res = IVar() in Ch.give (server_blockizer x.spiPath) (Req.Modify(x.spiEdits,res)) >>=. IVar.read res)
            |> run |> Json.serialize
        | FileTokenRange x ->
            (let res = IVar() in Ch.give (server_blockizer x.spiPath) (Req.GetRange(x.range,res)) >>=. IVar.read res)
            |> run |> Json.serialize
```

I mean, look at this. I am using a router socket here, but there is no difference between it and a response socket in this context. All the cases need to finish before the program can move on.

1pm. The way VS Code API is organized, it is assumed that all the providers are basically remote procedure calls. That means, they behave much like regular function calls in that the send some data, and then get a return. They aren't async message sends.

1:05pm. I could have more flexibility if I bring in the thread safe NetMQ queue.

That would make things complicated. Instead, let me simplify things to the limit.

```fs
use sock = new RouterSocket()
```

Why don't I in fact make this a response socket. Why not fully embrace the notion that the VS Code plugin (written in TS) and the F# language server are one whole and that the message sends are really remove function calls.

Then for the hover provider, why don't I put in a timeout of something like 0.25s. If typechecking takes longer than that, the VS Code side can just resend the message. Problem solved.

Using time to transmit information is really an elegant solution. I haven't really ever used it before.

Actually the same thing goes for parsing. Right now I am bothered how the range provider would need to wait for it to finish. Maybe the user might end up writing 1000 lines functions. I do not want him to wait until parsing is done to get semantic highlighting. I should be using timeouts for this as well.

Well, I'll deal with that particular issue later. Right now, I do not have any way to retrigger the range provider. I'll leave the parsing as it is.

1:20pm. The way it is currently structured is not a big problem.

...I am really getting strong deja vu as I imagine all of this. Suppose I separated out parsing errors from token errors. And suppose that I just send parser errors separately as soon as a particular block finishes.

Then I could easily implement parser timeouts. Even if I do not have a way to retrigger the range provider, it might not really matter since moving the viewport will cause it to do that.

1:25pm. Let me take a short break here.

1:45pm. Let me step away from the screen for a bit. Right now I am thinking about the parser. The way editor support works for it should be redone. The parser is one thing that should have proper timeouts.

I need to do it properly from the top. Then once I do that, I will be ready to tackle typechecking in a mock fashion. The same goes for hover info.

I need to think about this. Yeah, there is no doubt that if I were experienced in this, I would be able to do this much faster. In 2017, Spiral took me a sick amount of time. I can't get impatient with v0.2. It will take whatever it needs to take."