"9:55am. I am dead tired right now. Yesterday I went to bed earlier than usual, at 10pm, a time that can be considered regular. Then as I drifted in my thoughts I've been gripped by an intense feeling of longing.
I have my issues, and I have my regrets, but this feeling of longing made me think that I want my relationship to programming to be something more. One thing I am absolutely certain of is that I do not want to drag around this loser attitude.

Imagine if programming were wizardry.

A wizard would not regret not having money or respect, as long as he had arcane power. The feeling of studying magic would be enough of a reward for him.

That one day he might become powerful enough to become a god is enticing as a thought, but he does not dwell too much of that. Thinking about godhood is a distraction from the more rewarding everything day grind for power and understanding to him. If somebody came to him and showed him the future where he died prematurely to master the high tiers of spellcasting needed for that, he would not break and exchange studying magic for wordly pursuits like wine and women.

10:05am. I feel a longing to be like such a wizard in my imagination. I do not want to attach the importance of what I am doing because it will lead me to causing the Singularity or making money down the road.

I should be doing programming for its own sake. I should love its unbridled potential for its own sake.

Programming might not be wizardry (yet), but it is closest one will get to it in the real world. Somewhat ironically, it is closer to wizardry than wizardry itself, but it is also more fake. Today's games are fake, but I sometimes get echoes of reality in programming when investigating the limits of algorithms and languages. One day that echo will be reflected in highest forms of art as well.

10:15am. The effort I've put in the last 5.5 years, and the attitude I've effected has been nothing short of legendary. I did all the right things, for the wrong reason. I came (back) into programming from the other end, with the goals and intentions fully formed. I do owe a lot to those goals and the dream of Singualrity. But now I think it is time to accept that it would have been better if I were a programmer who discovered those things along the way, rather than a speculator changing classes midway to a programmer.

10:20am. It is the great irony that power is the reason for everything, but letting it define you is just arrogance.

I want to cultivate the ideal programmer's mindset. I do not want to have any weaknesses at all. I want my identity to be proper.

If I were a martial artist, it is fine that I spent 10 years living in the mountains tempering myself. No need to live in the ring getting punched and accumulating brain damage. It is okay to fight only when you've leved up enough. Once you have a high enough level, you get to the point where you are only dishing out damage and never taking it.

When you have those dynamics, the battles are a side note, and the life of self improvement is what is foremost and center.

10:25am. Seen for the past 5.5 years, my work ethic and focus might have been very high, but I should know the truth best of all. That I am a fake, and that I am going at this with the mindset of a degenrate gambler. I keep gambling my time. I keep putting 6 months on the table and never hitting my more ambitious goals.

10:30pm. The only thing I've ever succeeded at is getting better at programming.

Since that is my only success in life, I think I should accept my talent at it and make it my core. I do not want to bemoan my fate. I should not want the Inspired Desire just to do programming.

10:30am. Since I have such strong feelings today, how about I try to direct them into actual programming. I dealt with the typechecker and a new phase will start here.

As time goes on even the wizard will change."