"3:35pm. What I feel like doing now is reading LNs. Just forget about my responsibilities. Maybe I'll just go to bed earlier today? Who knows. My habits seem to be changing. Or am I changing them myself?
It is just, for the things I want to do, I want to actually have that feeling. Programming is a constant trip into darkness. When you make a few steps, it start surrounding you and the flame of desire gets suppressed.

4:50pm. Ok, let me finish the thing I am reading and I'll do a bit more. I got just a little bit of interest for doing some refactoring. I've realized that checking unbound vars and infer can be put into their own module. That is what I am going to do.

There is a lot of crap in `Typechecker.fs`, but I just need two functions (and some types) and I am gold.

4:55pm. Let me do just a bit more.

Part of the reason why I was so uncomfortable with starting is because the module is a mess. But if I can clean it up just a tad, I should be able to stoke my desire a little.

Right now everything feels coupled, but that is not true - `infer` is actually quite self contained.

5pm.

```fs
let validate_bound_vars (top_env : Env) term ty x =
```

Let me rename the template for calcing unbound vars. Rather than throw an exception this one just returns an array of them. That is the errors related to them.

5:05pm. Good, really good. Though I've mentioned that the typechecker is only 80% complete, I can reasonably make the claim that the stuff in `Infer.fs` is between 95-100% done. It is almost done. Later, I'll just be adding extra constraints and code for linking variables together for the sake of editor support. But those are just extras.

It is like having one big area, and having a mess on one side of it. All that I've done is put a fence around the garbage heap and now things feel a lot cleaner.

5:10pm. Now I can think more clearly.

The `Infer` module is 100% done effectively. And to complete the typechecker, all I have to do is take care of the top level - those `BundleTop` statements.

I already have code for doing half of them. All I have to do is take care of regular inl and mutually recursive blocks of them. And I am not sure whether I handled mutually recursive types properly either.

At any rate, half seems like a decent estimate and what I have now is 70 LOC.

5:15pm. Now that I've though this, I can visualize the path twoards finishing the typechecker a lot more clearly. If not finish it, I should be able to make a big dent tomorrow.

The work I've done on editor support in the past week did not really register in my mind as making progress towards finishing the TC, but now I can see it. The gap has definitely narrowed. I should be able to move to testing before the month is over.

5:20pm. I feel completely fine finishing the day with this.

Yeah, the way I behaved today was just fine. It should be fine to trust myself a bit more. It is fine to not work if you do not feel like it. I do not have to be like a cop watching over myself. I won't ever let myself fall to indulgence. And I will trust myself to maintain a balance."