"9:45am. Another day is here, let me slack for a while. I am groggy.
Somehow I am really enjoying myself reading LNs lately. This kind of life is fine.

https://www.080910t.com/downloads/AI%20Chip%202018%20EN.pdf
White Paper on AI Chip Technologies

A person I am in contact this linked me this.

10am. Mhhhh...Since I have only a few chapters left, let me catch up with the vamp loli paladin story, then I'll read the above white paper. After that come breakfast. After that I think I'll tap into the desire I established yesterday and start work on the top level statements.

...Actually, let me just have breakfast now.

11:20am. Done with breakfast. Let me slack for just a little more and then I'll go through that AI chip paper. Then comes programming. I am really sick of the typechecker, so I want to get it out of the way. I am not at all in the mood to fiddle around with it going into the future. I'll deal with it, and move on.

11:45am. Finally done. Let me do some work for the day.

First come the white paper. Actually, I completely forgot that in addition to `infer` I've still yet to do the function that annotates the AST and validates that there are no lingering metavars.

So today my focus will be on the top level, but I'll have to complete the typed AST transform before I go to the next part.

11:50am. Well, let me take it slowly. Forget distant goals. Just focus on the next step. It is best to take it one at a time. My mission for the day will be to just get into it and let it come to me on its own.

>  The flexibility of this powerful processing power, is also reflected in the training and inference task deployment, such as more hardware resources are allocated to inference tasks during the day time, while these resources are allocated to training tasks at night.

Day = inference. Night = training. It is interesting to think of it that way. Sleep is just such a huge disadvantage for a living being, that nature had to include it for some reason. It makes sense, but I need to keep in mind that this is still just speculation. How would one prove something like that?

TODO: I want to have the let statements be non boxing. I got the urge to do this yesterday, and I still have it so I'll go for it.

12:40pm.

```fs
        | false, _, [], [] ->
            match patterns_validate [name] with
            | [] -> Ok((r,dyn_if_let name,body),is_rec)
            | ers -> Error ers
```

I just have to get rid of that `dyn_if_let` here no problem. Let me get back to the paper.

1:25pm. Had to take a little break. I am done with the paper.

Since I've been following tech development on and off in this area since the early 10s, I don't have much to comment on. I was a nice refresher.

1:30pm. Let me take another short break and then I will start programming for the day."